Updated the implementation for strategy design
Updated the implementation of the strategy design pattern with the client-maintenance component
