--- client/src/business-layer/ClientMaintenance/CreateAgreement.ts
@@ -1,9 +1,0 @@
-class CreateAgreementC1 {
-
-}
-
-function CreateAgreement() {
-    
-}
-
-export default CreateAgreement;
