--- client/src/business-layer/ClientMaintenance/TerminateAgreement.ts
@@ -1,9 +1,0 @@
-class TerminateAgreementC1 {
-
-}
-
-function TerminateAgreement() {
-    
-}
-
-export default TerminateAgreement;
