--- client/src/business-layer/ClientMaintenance/UpdateAgreement.ts
@@ -1,9 +1,0 @@
-class UpdateAgreementC1 {
-
-}
-
-function UpdateAgreement() {
-    
-}
-
-export default UpdateAgreement;
