+++ client/src/business-layer/client-maintainance/IndividualNotes.ts
