+++ client/src/business-layer/client-maintainance/NoteRecorder.ts
