+++ client/src/business-layer/client-maintainance/ServiceAgreementStrategy.ts
@@ -1,0 +1,6 @@
+import ServiceAgreement from "../../data-layer/data-classes/ServiceAgreement";
+
+interface ServiceAgreementStrategy {
+    HandleAgreement(serviceAgreement: ServiceAgreement): void;
+  }
+export default ServiceAgreementStrategy;
