+++ client/src/business-layer/client-maintainance/TerminateAgreement.ts
@@ -1,0 +1,12 @@
+import ServiceAgreement from "../../data-layer/data-classes/ServiceAgreement";
+import ServiceAgreementStrategy from "./ServiceAgreementStrategy";
+
+class TerminateAgreementC1 implements ServiceAgreementStrategy {
+
+    HandleAgreement(serviceAgreement: ServiceAgreement): void {
+        //Logic for teminating an Agreement
+    }
+
+}
+
+export default TerminateAgreementC1;
