+++ client/src/business-layer/client-maintainance/UpdateAgreement.ts
@@ -1,0 +1,12 @@
+import ServiceAgreement from "../../data-layer/data-classes/ServiceAgreement";
+import ServiceAgreementStrategy from "./ServiceAgreementStrategy";
+
+class UpdateAgreementC1 implements ServiceAgreementStrategy {
+
+    HandleAgreement(serviceAgreement: ServiceAgreement): void {
+        //Logic for updating an Agreement
+    }
+
+}
+
+export default UpdateAgreementC1;
